+++ types/drawio/index.d.ts
@@ -3,6 +3,8 @@
 declare namespace Draw {
   function loadPlugin(plugin: ((ui: UI) => any)): void;
 
+  type DatabaseItem = {key: any, data: any};
+
   class UI {
     actions: Actions;
     currentPage: DiagramPage;
@@ -19,10 +21,27 @@ declare namespace Draw {
     fileNode: FileNode;
     editor: Editor;
     toolbar: Toolbar;
-    showDialog(elt, w, h, modal, closable, onClose?, noScroll?, transparent?, onResize?, ignoreBgClick?);
+    showDialog(
+      element: HTMLElement,
+      width: number,
+      height: number,
+      modal: boolean, // Add background?
+      closable: boolean, // Show close icon?
+      onClose: (cancel?: boolean, isEsc?: boolean) => boolean | void, // Return value specifies whether dialog shall be closed
+      noScroll?: boolean, // Dialog not scrollable?
+      transparent?: boolean, // Dialog transparent?
+      onResize?: () => void,
+      ignoreBgClick?: boolean
+    ): void;
     confirm(msg: string, okFn: () => void, cancelFn?: () => void, okLabel?: string, cancelLabel?: string, closable?: boolean): void;
-    hideDialog();
+    hideDialog(cancel?: boolean, isEsc?: boolean, matchContainer?: HTMLElement): void;
     removeLibrarySidebar(id: string);
+
+    // IndexDB operations
+    setDatabaseItem(key: any, data: any, success: (() => void) | null, error: (() => void) | null, storeName?: string): void;
+    removeDatabaseItem(key: any, success: (() => void) | null, error: (() => void) | null, storeName?: string): void;
+    getDatabaseItem(key: any, success: ((result?: DatabaseItem) => void) | null, error: (() => void) | null, storeName?: string): void;
+    getDatabaseItems(success: ((result?: DatabaseItem[]) => void) | null, error: (() => void) | null, storeName?: string): void;
   }
 
   class DiagramPage {
@@ -172,6 +191,21 @@ class mxChildChange {
 }
 
 class mxEventObject { }
+
+type WorkerRequestTypes = 'Function' | 'Release';
+
+interface WorkerRequest {
+  readonly type: WorkerRequestTypes,
+}
+interface WorkerFunctionRequest extends WorkerRequest {
+  readonly fn: string,
+  readonly data: unknown,
+}
+
+interface WorkerResponse {
+  result?: unknown,
+  error?: string,
+}
 
 // Globally available mxgraph exports
 const mxChildChange: typeof import('mxgraph').mxChildChange;